Constraints Overhaul Divided the God-Class ConstraintVariable into smaller, more self governing classes. ConstraintTree handles overall computation and is comparible to the previous Constraint class  - handles all ! symbols  - handles all brackets  - builds itself from the addElement method  - primary attribute is the root ConstraintVariable
ConstraintOperator handles operators and the conversion from their normal form to their negated form
 - handles ConstraintNumberLiterals and puts them to the right of the operator
 - handles ConstraintLiterals(strings) and puts them on the left of the operator (==)
 - handles the conversion of the == or equals methods
 - handles making the association produce a "numberOf" method(instead of a .size() method)

ConstraintNamed encompasses a class of ConstraintVariable classes with the following characteristics
 - they represent 1 variable, an attribute for example, or state
 - the variable they represent has a name
 - the varieties of ConstraintNamed are: ConstraintAttribute, ConstraintAssociation, ConstraintUnassignedName, ConstraintMethodParameter, ConstraintState, and ConstraintStateMachine
 - these to little to no computation, and are only used to differeniate between the different classes

ConstraintLiteral handles values that are immutable, such as strings and numbers
 - there is a subclass for numbers, ConstraintNumberLiteral

ConstraintVariable is the super class, and ensures three things
 1. there is a getType() method that returns a generic type, such as "attribute", or "operatorTwo" which helps the SuperGenerator distinguish which method to use when translating the ConstraintVariable
 2. there is a getValue(CodeTranslator) method for every ConstraintVariable, which uses the getType method, and the translate(String,ConstraintTree) method.
 3. there is a iterator() method.

All ConstraintVariables have a iterator<ConstraintVariable> associated to them, ConstraintLiteral and ConstraintNamed only have one value to iterate over, so they trivially return themselves and are done, this is the default.
 - ConstraintTree, which is usually the root of the constraint proper, iterates over all of its root's iterator's values before returning itself
 - ConstraintOperator iterates over its left ConstraintVariable, then return's itself, then returns its right, excluding the left if it doesn't exist
The iterators can be used to iterate over the whole of the tree. for example:
for(ConstraintVariable variable:tree)
{
}

This change facilitates some powerful differences, mainly that the getMethod's of the attributes(etc.) are now used instead of their simple names.
That means that inheritance could now work if implemented. All that would be required would be a method that overrides the setMethod for the attribute with the constraint, and a call to the superclass's version with the if statement generated by the constraint
for example
class X{int age;}
class Y{isA X; int weight; [age>18&&weight<3]}

would have a method in class Y
public boolean setWeight(int aWeight)
{
  if(getAge()>18&&aWeight<3)
  {
    return super.setWeight(aWeight);
  }
  else return false;
}
(note: this hasn't been implemented yet, however this refactoring provides that next step, whereas before such a thing would be impossible)

Two miscillaneous changes needed to be made to make this new ConstraintTree system work
 1. key's can now accept derived attributes and will not try to use the simple name of the key, for example if you have: variable={"value"}; key{variable}. The equals function will contain if(compareTo.equals(getVariable()) instead of if(compareTo.equals(variable))
 2. in order to accept templated inheritance, quotes can be used to exactly specify the class, so isA "Iterable<ConstraintVariable>"; will generate the implements Iterable<ConstraintVariable> as desired. The quotes are partial necessary because of how the traits are configured.



git-svn-id: http://umple.googlecode.com/svn/trunk@4558 6ebf9d09-8eb1-0036-9d08-68c07d1b990a